Move and rename sparsifyLocation to var-model.js
